Added Authentication in testig Category
